[api] Add queues system
The idea of this queue system is we start by prefill the table
with available stocks and when user want to reserve the vaccine
we basically get a queue from the table and set it status as reserving.

Benefit of this approach is we don't need to worry about phantom write
and we only need to lock the row that we really need. as 👇 example

```sql
UPDATE queues
    SET token = <new random token>, expired_at = <new expired date>
    WHERE id = (
            SELECT q.id FROM queues q
            WHERE <not paid> and <expired>
            FOR UPDATE SKIP LOCKED
         )
```